[REEF-81]: Rename com.example to org.apache.reef.tang.examples
JIRA: [REEF-81](https://issues.apache.org/jira/browse/REEF-81)
