+++ reef-tang/tang-test-jarA/src/main/java/org/apache/reef/tang/examples/A.java
@@ -16,8 +16,8 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.example;
+package org.apache.reef.tang.examples;
 
-public class B extends A {
-
+public class A {
+  int iAmAClass = 0;
 }
