+++ reef-tang/tang-test-jarAB/src/main/java/org/apache/reef/tang/examples/A.java
@@ -16,10 +16,10 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.example;
+package org.apache.reef.tang.examples;
 
 import javax.inject.Inject;
 
-public class B extends A {
-  @Inject public B() { }
+public class A {
+  @Inject public A() {}
 }
