+++ reef-tang/tang-test-jarAB/src/main/java/org/apache/reef/tang/examples/B.java
@@ -16,8 +16,10 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.example;
+package org.apache.reef.tang.examples;
 
-public class A {
-  int iAmAClass = 0;
+import javax.inject.Inject;
+
+public class B extends A {
+  @Inject public B() { }
 }
